Adding the rest Responses post and get

+++ src/main/java/ca/gc/esdc/springtestproject/dao/Repository/ProjectRepository.java
@@ -3,6 +3,6 @@ package ca.gc.esdc.springtestproject.dao.Repository;
 import ca.gc.esdc.springtestproject.dao.entities.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjectRepository extends JpaRepository<Project,Integer>  {
+public interface ProjectRepository extends JpaRepository<Project,Long>  {
 
 }
